requirements.txt
pylint
django
djangorestframework
djangorestframework-simplejwt
django-rest-knox
pymysql
psycopg2-binary
django-cors-headers
django-rest-auth
pillow
glob2
drf-jwt
PyJWT
requests
pylint
python-decouple
black

asgiref==3.4.1
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5